Add files via upload
This is a direct update for the main branch. No forks. 
We found two issues. One on the ability to select an operational mode. The second to address output generation. 
At this time only minimal has been fully tested. Still testing on Magic and Multi. 
When complete Magic will become an integrated option for MultiMode allowing the operator to choose a motion priority handler, Software, Firmware, or Magic.