Add files via upload
Wrap up of several fixed and unfortunately even replaced the old files with mostly rewritten files as the first 3 versions were apparently just for testing... Nothing like arguing with an AI. Like, they were always supposed to be functional. I dont mind having one for testing and verification that the code works but it is supposed to all generate code!!!
So anyway, now they all work. Not to the point of celebration. But they generate code that works fine on Marlin. I expect to keep enhancing this to until its something really nice. But minimal will always be as basic as I can keep it. Just filters and simple replacement of code as needed, maybe baseline machine parameters to keep the bus on the rails. 

What was fixed? We made it so the code generate right. Exports right.
What was added? We added headers, start up code and shutdown code. 
These are intended to be functional as switches so you dont have to use the, dont have to add them to your files. 
We added the ability to include start up and shout down scripts. They are not considered by the post processor and will be passed on directly in the output code. 
You have the option to post process the file as a NC or gcode. This only effects the file name.