Add files via upload
With the post processors working. I have added the ability to deal with reversed axis. Most of us should have flushed these issues out in the development phase of our hardware others have no control over that. In theory Fusion360 also offers tools to deal with this but in the end, its simplest to just leave everything in the natural state and use a tool like a post processor to set the world right. So your welcome, now we can reverse axis, in the post processor, no other tinkering required. -Note Github quietly tried to bump copilot to chat gpt 5 mini, which is supposed to be faster but was way slower and I have not well tested. Actually I liked 4.0 but it's no longer available. So what this means is that there may be unexpected issues in this update. I will flush as quickly as I can, if found. But of course, users are also welcome to help out. -This update also includes slightly expanded TMC support in the multi mode. It should already be there as of one commit prior but if not it is now. Again slightly more support. The point is to flush out abilities that normaly sleep or get over looked to create a better experience. Not create a whole new story. Marlin is still in charge. We seek only to expose all it's features and make as best use of them as we can.